@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "run-s -c lint:ts lint:next lint:prettier",
@@ -18,24 +18,25 @@
   "dependencies": {
     "@serwist/next": "^9.0.11",
     "bootstrap": "^5.3.3",
-    "next": "^14.2.23",
-    "react": "^18.3.1",
+    "next": "^15.1.4",
+    "react": "^19.0.0",
     "react-bootstrap": "^2.10.7",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "react-use-audio-player": "^3.0.2",
     "sharp": "^0.33.5"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.17.0",
     "@types/bootstrap": "^5.2.10",
     "@types/eslint-config-prettier": "^6.11.3",
+    "@types/eslint__eslintrc": "^2.1.2",
     "@types/lint-staged": "^13.3.0",
     "@types/node": "^22.10.5",
-    "@types/react": "^18.3.18",
-    "@types/react-dom": "^18.3.5",
-    "@typescript-eslint/eslint-plugin": "^8.19.1",
-    "@typescript-eslint/parser": "^8.19.1",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "^14.2.23",
+    "@types/react": "^19.0.4",
+    "@types/react-dom": "^19.0.2",
+    "eslint": "^9.17.0",
+    "eslint-config-next": "^15.1.4",
     "eslint-config-prettier": "^9.1.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.3.0",
@@ -43,7 +44,8 @@
     "prettier": "^3.4.2",
     "prettier-plugin-packagejson": "^2.5.6",
     "serwist": "^9.0.11",
-    "typescript": "^5.7.3"
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.19.1"
   },
   "packageManager": "yarn@1.22.22"
 }
